:refactor: global domain추가
