--- src/main/java/oneclass/oneclass/auth/dto/FindUsernameRequest.java
@@ -1,10 +1,0 @@
-package oneclass.oneclass.auth.dto;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class FindUsernameRequest {
-    private String emailOrPhone;
-}
