--- src/main/java/oneclass/oneclass/auth/dto/LoginRequest.java
@@ -1,11 +1,0 @@
-package oneclass.oneclass.auth.dto;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class LoginRequest {
-    private String username;
-    private String password;
-}
