--- src/main/java/oneclass/oneclass/auth/dto/ResetPasswordRequest.java
@@ -1,12 +1,0 @@
-package oneclass.oneclass.auth.dto;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class ResetPasswordRequest {
-    private String username;
-    private String newPassword;
-    private String verificationCode;
-}
