--- src/main/java/oneclass/oneclass/auth/dto/ResponseToken.java
@@ -1,11 +1,0 @@
-package oneclass.oneclass.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ResponseToken {
-    private String accessToken;
-    private String refreshToken;
-}
