--- src/main/java/oneclass/oneclass/auth/entity/RefreshToken.java
@@ -1,24 +1,0 @@
-package oneclass.oneclass.auth.entity;
-
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class RefreshToken {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String username;//토큰 소유자
-    private String token;//refreshToken값
-
-    private LocalDateTime expiryDate;//만료일
-}
