--- src/main/java/oneclass/oneclass/auth/entity/Role.java
@@ -1,9 +1,0 @@
-package oneclass.oneclass.auth.entity;
-
-
-
-public enum Role {
-    STUDENT,
-    PARENT,
-    ADMIN
-}
