--- src/main/java/oneclass/oneclass/auth/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package oneclass.oneclass.auth.repository;
-
-import oneclass.oneclass.auth.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByUsername(String username);
-    Optional<Member> findByEmailOrPhone(String email, String phone);
-}
