--- src/main/java/oneclass/oneclass/auth/repository/RefreshTokenRepository.java
@@ -1,11 +1,0 @@
-package oneclass.oneclass.auth.repository;
-
-import oneclass.oneclass.auth.entity.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    Optional<RefreshToken> findByUsername(String username);
-    void deleteByUsername(String username);
-}
