--- src/main/java/oneclass/oneclass/auth/repository/VerificationCodeRepository.java
@@ -1,7 +1,0 @@
-package oneclass.oneclass.auth.repository;
-
-import oneclass.oneclass.auth.entity.VerificationCode;
-import org.springframework.data.repository.CrudRepository;
-
-public interface VerificationCodeRepository extends CrudRepository<VerificationCode, String> {
-}
