+++ src/main/java/oneclass/oneclass/domain/counsel/dto/ConsultationDetailResponse.java
@@ -1,0 +1,28 @@
+package oneclass.oneclass.domain.counsel.dto;
+
+import lombok.Data;
+import oneclass.oneclass.domain.counsel.entity.Consultation;
+
+@Data
+public class ConsultationDetailResponse {
+    private String name;
+    private String phone;
+    private String type;
+    private String subject;
+    private String description;
+    private String status;
+    private String scheduleTime;
+
+    // 생성자 혹은 static from(Entity) 메서드
+    public static ConsultationDetailResponse from(Consultation consultation) {
+        ConsultationDetailResponse res = new ConsultationDetailResponse();
+        res.setName(consultation.getName());
+        res.setPhone(consultation.getPhone());
+        res.setType(consultation.getType());
+        res.setSubject(consultation.getSubject());
+        res.setDescription(consultation.getDescription());
+        res.setStatus(consultation.getStatus());
+        // 필요하다면 scheduleTime 등도 추가
+        return res;
+    }
+}
