+++ src/main/java/oneclass/oneclass/global/auth/dto/FindUsernameRequest.java
@@ -1,0 +1,10 @@
+package oneclass.oneclass.global.auth.dto;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class FindUsernameRequest {
+    private String emailOrPhone;
+}
