+++ src/main/java/oneclass/oneclass/global/auth/dto/LoginRequest.java
@@ -1,0 +1,11 @@
+package oneclass.oneclass.global.auth.dto;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class LoginRequest {
+    private String username;
+    private String password;
+}
