+++ src/main/java/oneclass/oneclass/global/auth/dto/ResetPasswordRequest.java
@@ -1,0 +1,12 @@
+package oneclass.oneclass.global.auth.dto;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class ResetPasswordRequest {
+    private String username;
+    private String newPassword;
+    private String verificationCode;
+}
