+++ src/main/java/oneclass/oneclass/global/auth/dto/ResponseToken.java
@@ -1,0 +1,11 @@
+package oneclass.oneclass.global.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ResponseToken {
+    private String accessToken;
+    private String refreshToken;
+}
