+++ src/main/java/oneclass/oneclass/global/auth/entity/Member.java
@@ -1,0 +1,27 @@
+package oneclass.oneclass.global.auth.entity;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Member {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, unique = true)
+    private String username;//이름
+    private String password;//비번
+    private String phone;
+    private String email;
+    @Enumerated(EnumType.ORDINAL)
+    @NotNull
+    private Role role;
+}
