+++ src/main/java/oneclass/oneclass/global/auth/entity/RefreshToken.java
@@ -1,0 +1,24 @@
+package oneclass.oneclass.global.auth.entity;
+
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class RefreshToken {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String username;//토큰 소유자
+    private String token;//refreshToken값
+
+    private LocalDateTime expiryDate;//만료일
+}
