+++ src/main/java/oneclass/oneclass/global/auth/entity/Role.java
@@ -1,0 +1,9 @@
+package oneclass.oneclass.global.auth.entity;
+
+
+
+public enum Role {
+    STUDENT,
+    PARENT,
+    ADMIN
+}
