+++ src/main/java/oneclass/oneclass/global/auth/entity/VerificationCode.java
@@ -1,0 +1,19 @@
+package oneclass.oneclass.global.auth.entity;
+
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import lombok.*;
+
+@Entity
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class VerificationCode {
+    @Id
+    private String usernameOrEmail;
+    private String code;
+    private Long expiry;
+}
