+++ src/main/java/oneclass/oneclass/global/auth/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package oneclass.oneclass.global.auth.repository;
+
+import oneclass.oneclass.global.auth.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByUsername(String username);
+    Optional<Member> findByEmailOrPhone(String email, String phone);
+}
