+++ src/main/java/oneclass/oneclass/global/auth/repository/RefreshTokenRepository.java
@@ -1,0 +1,11 @@
+package oneclass.oneclass.global.auth.repository;
+
+import oneclass.oneclass.global.auth.entity.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+    Optional<RefreshToken> findByUsername(String username);
+    void deleteByUsername(String username);
+}
