+++ src/main/java/oneclass/oneclass/global/auth/repository/VerificationCodeRepository.java
@@ -1,0 +1,7 @@
+package oneclass.oneclass.global.auth.repository;
+
+import oneclass.oneclass.global.auth.entity.VerificationCode;
+import org.springframework.data.repository.CrudRepository;
+
+public interface VerificationCodeRepository extends CrudRepository<VerificationCode, String> {
+}
